% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdmtools.R
\docType{_PACKAGE}
\name{gdmtools}
\alias{gdmtools}
\title{GDM tools: post-processing GDM output rasters}
\description{
Providing functions and utilities to generate the advanced spatial analysis
 products from GDM raster outputs (based on Mokany et al. 2022).
}
\references{
rvalavi/gdmtools, Mokany et al. 2022...
}
\seealso{
\code{\link{par_index}}
}
\author{
Roozbeh Valavi, etc..
}
