% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_index.R
\name{par_index}
\alias{par_index}
\title{Protected area representativeness}
\usage{
par_index(x, pa, mod, n = NULL, ncores = -1, ...)
}
\arguments{
\item{x}{A SpatRaster of GDM transformed layers. It accepts any raster type convertible to terra objects
or a path to files on disk.}

\item{pa}{protected areas as polygons, masked GDM transformed rasters, or a data.frame of extracted values
from protected areas.}

\item{mod}{either a gdm model or the intercept of fitted model.}

\item{n}{number of samples to take from the entire region to normalise the similarity values. if NULL or 0
used, 20% of the number of non-na cells will be used as the sample number.}

\item{ncores}{number of cores to process. Default is all available cores.}

\item{...}{Additional arguments for writing raster outputs e.g. \code{filename},
\code{overwrite}, and \code{wopt} from terra \code{\link[terra]{predict}}.}
}
\value{
SpatRaster
}
\description{
Using GDM predictions for assessing how well the protected areas in a
region represent the biological diversity.
}
\examples{
library(gdmtools)

}
