% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persistance.R
\name{persistance}
\alias{persistance}
\title{Estimate Expected Species Persistence based on Habitat Condition}
\usage{
persistance(x, y, mod, power = 0.25, n = NULL, ncores = -1, ...)
}
\arguments{
\item{x}{A SpatRaster of GDM-transformed layers. Accepts any raster type that can be converted
to a \pkg{terra} object or a file path to rasters stored on disk.}

\item{y}{A SpatRaster of habitat condition. Accepts any raster type that can be converted
to a \pkg{terra} object or a file path to rasters stored on disk.}

\item{mod}{Either a \pkg{gdm} model object or the intercept value from a fitted model.}

\item{power}{Numeric. The power of z for species-area conversion (default 0.25).}

\item{n}{Numeric. Specifies the number of samples to take from the GDM transformed layers
and habitat condition data. If \code{NULL} or 0, a sample size of 10\% of the non-NA cells
will be used. This random sample may include NA values, which will be removed. Consequently,
the final number of samples used for analysis may be lower than the specified sample size.}

\item{ncores}{Integer. Specifies the number of CPU threads to be used for processing. A value
below 1 indicates that all available threads will be utilized. Refer to the details section for
more information.}

\item{...}{Additional arguments for writing raster outputs e.g. \code{filename},
\code{overwrite}, and \code{wopt} from terra \code{\link[terra]{predict}}.}
}
\value{
SpatRaster
}
\description{
This function combines spatial predictions from Generalized Dissimilarity
Modelling (GDM) with habitat condition data to estimate the expected level
of species persistence at each location.
}
\examples{
# library(gdm)
library(gdmtools)

}
\references{
Mokany, K., Ware, C., Woolley, S. N. C., Ferrier, S., & Fitzpatrick, M. C. (2022).
A working guide to harnessing generalized dissimilarity modelling for biodiversity analysis and
conservation assessment. Global Ecology and Biogeography, 31(4), 802–821.
}
